backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# File handling
python-multipart==0.0.6
python-docx==1.1.0

# Language processing
language-tool-python==2.7.1
sentence-transformers==2.2.2

# Hugging Face Transformers for LLM
transformers==4.35.0
torch==2.1.0
accelerate==0.24.0
tokenizers>=0.14,<0.15

# HTTP client for external APIs
httpx==0.25.2
aiohttp==3.9.1

# Machine learning and data processing
numpy==1.24.3
scikit-learn==1.3.2

# Text processing
nltk==3.8.1
spacy==3.7.2

# Authentication and Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Email
jinja2==3.1.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Development tools
python-dotenv==1.0.0 

# Database
SQLAlchemy>=2.0.0
asyncpg>=0.28.0
alembic>=1.13.0
psycopg[binary]>=3.1.0
aiosqlite>=0.19.0